Lab 14 - Implement the update-scopes-generator
We want to make sure that scope tags specified in our util-lib
generator are up to date and take into account all these new scopes
that teams have been adding.

We want to check if there is a new scope tag in any of our project.json
files and update our generator schema.